replace test_glm with test_glm_2
